Move logging and database management to separate modules.
