+++ src/wpp/RunReports.py
@@ -3,11 +3,10 @@ import argparse
 import datetime as dt
 import sqlite3
 import pandas as pd
-import logging
 import copy
 import sys
 import os
-from typing import Optional, Tuple, Any, cast
+from typing import Tuple, cast
 
 from wpp.config import (
     WPP_REPORT_DIR,
@@ -16,40 +15,12 @@ from wpp.config import (
     WPP_REPORT_FILE,
 )
 from wpp.calendars import BUSINESS_DAY
+from wpp.db import get_single_value, run_sql_query, union_sql_queries, join_sql_queries
+from wpp.logger import get_log_file
 
-WPP_LOG_FILE = WPP_LOG_DIR + r"/Log_RunReports_{}.txt"
-
-
-#
-# Set up Logging
-#
-class STDOutFilter(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        return record.levelno >= logging.INFO  # | record.levelno == logging.DEBUG
-
-
-class STDErrFilter(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        return not record.levelno == logging.INFO | record.levelno == logging.DEBUG
-
-
-today = dt.datetime.today()
-os.makedirs(WPP_LOG_DIR, exist_ok=True)
-log_file = WPP_LOG_FILE.format(today.strftime("%Y-%m-%d"))
-logFormatter = logging.Formatter(
-    "%(asctime)s - %(levelname)s: - %(message)s", "%H:%M:%S"
-)
-logger = logging.getLogger()
-handler = logging.FileHandler(log_file)
-handler.setFormatter(logFormatter)
-logger.addHandler(handler)
-
-handler = logging.StreamHandler()
-handler.setFormatter(logFormatter)
-handler.addFilter(STDOutFilter())
-logger.addHandler(handler)
-
-logger.setLevel(logging.INFO)
+# Set up logger
+LOG_FILE = WPP_LOG_DIR + r"/Log_RunReports_{}.txt"
+logger = get_log_file(__name__, LOG_FILE)
 
 #
 # SQL
@@ -317,67 +288,6 @@ ORDER BY block_ref
 """
 
 
-def get_db_connection(db_file: str) -> sqlite3.Connection:
-    conn = sqlite3.connect(db_file)
-    return conn
-
-
-def join_sql_queries(query_sql: str, sql1: str, sql2: str) -> str:
-    sql1 = sql1.replace(";", "")
-    sql2 = sql2.replace(";", "")
-
-    sql = query_sql.format(sql1, sql2)
-    return sql
-
-
-def union_sql_queries(
-    sql1: str, sql2: str, order_by_clause: Optional[str] = None
-) -> str:
-    sql1 = sql1.replace(";", "")
-    sql2 = sql2.replace(";", "")
-
-    sql = """
-    {}
-    UNION ALL
-    {}
-    """.format(sql1, sql2)
-    if order_by_clause:
-        sql += " " + order_by_clause
-    sql += ";"
-    return sql
-
-
-def run_sql_query(
-    db_conn: sqlite3.Connection, sql: str, args_tuple: Tuple
-) -> pd.DataFrame:
-    try:
-        df = pd.read_sql_query(sql, db_conn, params=args_tuple)
-        return df
-    except db_conn.Error as err:
-        logging.error(str(err))
-        # traceback.print_tb(ex.__traceback__)
-        logging.exception(err)
-        logging.error("The SQL that caused the failure is:")
-        logging.error(sql)
-        raise
-    except Exception as ex:
-        logging.error(str(ex))
-        # traceback.print_tb(ex.__traceback__)
-        logging.exception(ex)
-        raise
-
-
-def get_single_value(
-    db_cursor: sqlite3.Cursor, sql: str, args_tuple: Tuple = ()
-) -> Optional[Any]:
-    db_cursor.execute(sql, args_tuple)
-    value = db_cursor.fetchone()
-    if value:
-        return value[0]
-    else:
-        return None
-
-
 def add_column_totals(df: pd.DataFrame) -> pd.DataFrame:
     if len(df) > 0:
         # df = df.append(df.sum(numeric_only=True).rename("Total"))
@@ -432,19 +342,19 @@ def checkDataIsPresent(
     csr = db_conn.cursor()
     sql = "select count(ID) from Transactions where pay_date = ?"
     count: int = cast(int, get_single_value(csr, sql, (bos_date,)))
-    logging.info(f"{count} Bank Of Scotland transactions found for date {bos_date}")
+    logger.info(f"{count} Bank Of Scotland transactions found for date {bos_date}")
     is_data_present = is_data_present and (count > 0)
 
     sql = "select count(ID) from AccountBalances where at_date = ?"
     count = cast(int, get_single_value(csr, sql, (bos_date,)))
-    logging.info(
+    logger.info(
         f"{count} Bank Of Scotland account balance records found for date {bos_date}"
     )
     is_data_present = is_data_present and (count > 0)
 
     sql = "select count(ID) from Charges where at_date = ?"
     count = cast(int, get_single_value(csr, sql, (qube_date,)))
-    logging.info(f"{count} Qube charge records found for date {qube_date}")
+    logger.info(f"{count} Qube charge records found for date {qube_date}")
     is_data_present = is_data_present and (count > 0)
     return is_data_present
 
@@ -459,29 +369,29 @@ def runReports(db_conn: sqlite3.Connection, qube_date: str, bos_date: str) -> No
     # start_date = "{}-{}-{}".format(year, month, "1")
     # end_date = "{}-{}-{}".format(year, month, last_day_of_month)
 
-    logging.info(f"Qube Date: {qube_date}")
-    logging.info(f"Bank Of Scotland Transactions and Account Balances Date: {bos_date}")
+    logger.info(f"Qube Date: {qube_date}")
+    logger.info(f"Bank Of Scotland Transactions and Account Balances Date: {bos_date}")
 
     if not checkDataIsPresent(db_conn, qube_date, bos_date):
-        logging.error(
+        logger.error(
             f"The required data is not in the database. Unable to run the reports for Qube date {qube_date} and BoS transactions date {bos_date}"
         )
         sys.exit(1)
 
     # Create a Pandas Excel writer using openpyxl as the engine.
     excel_report_file = WPP_REPORT_FILE.format(qube_date)
-    logging.info(f"Creating Excel spreadsheet report file {excel_report_file}")
+    logger.info(f"Creating Excel spreadsheet report file {excel_report_file}")
     excel_writer = pd.ExcelWriter(excel_report_file, engine="openpyxl")
 
     # Run SC total transactions by block (COMREC) report for given run date
-    logging.info(f"Running COMREC report for {bos_date}")
+    logger.info(f"Running COMREC report for {bos_date}")
     sql = union_sql_queries(
         SELECT_TOTAL_PAID_SC_BY_PROPERTY_SQL,
         SELECT_TOTAL_PAID_SC_BY_BLOCK_SQL,
         "ORDER BY Reference",
     )
-    logging.debug(sql)
-    df = run_sql_query(db_conn, sql, (bos_date,) * 4)
+    logger.debug(sql)
+    df = run_sql_query(db_conn, sql, (bos_date,) * 4, logger)
     df = add_column_totals(df)
     df.to_excel(
         excel_writer,
@@ -489,28 +399,28 @@ def runReports(db_conn: sqlite3.Connection, qube_date: str, bos_date: str) -> No
         index=False,
         float_format="%.2f",
     )
-    df = run_sql_query(db_conn, BLOCKS_NOT_IN_COMREC_REPORT, (bos_date,) * 2)
+    df = run_sql_query(db_conn, BLOCKS_NOT_IN_COMREC_REPORT, (bos_date,) * 2, logger)
     blocks = df["Block"].tolist()
     if len(blocks) > 0:
-        logging.info(
+        logger.info(
             "Blocks which have transactions but are missing from the COMREC report because there is no bank account for that block: {}".format(
                 ", ".join(blocks)
             )
         )
 
     # Non-DC/PAY type transactions
-    logging.info(f"Running Transactions report for {bos_date}")
-    logging.debug(SELECT_NON_PAY_TYPE_TRANSACTIONS)
-    df = run_sql_query(db_conn, SELECT_NON_PAY_TYPE_TRANSACTIONS, (bos_date,) * 2)
+    logger.info(f"Running Transactions report for {bos_date}")
+    logger.debug(SELECT_NON_PAY_TYPE_TRANSACTIONS)
+    df = run_sql_query(db_conn, SELECT_NON_PAY_TYPE_TRANSACTIONS, (bos_date,) * 2, logger)
     df = add_column_totals(df)
     df.to_excel(
         excel_writer, sheet_name="Transactions", index=False, float_format="%.2f"
     )
 
     # DC/PAY type transactions
-    logging.info(f"Running DC & PAY Transactions report for {bos_date}")
-    logging.debug(SELECT_PAY_TYPE_TRANSACTIONS)
-    df = run_sql_query(db_conn, SELECT_PAY_TYPE_TRANSACTIONS, (bos_date,) * 2)
+    logger.info(f"Running DC & PAY Transactions report for {bos_date}")
+    logger.debug(SELECT_PAY_TYPE_TRANSACTIONS)
+    df = run_sql_query(db_conn, SELECT_PAY_TYPE_TRANSACTIONS, (bos_date,) * 2, logger)
     df = add_column_totals(df)
     df.to_excel(
         excel_writer,
@@ -520,14 +430,14 @@ def runReports(db_conn: sqlite3.Connection, qube_date: str, bos_date: str) -> No
     )
 
     # Run Qube BOS By Block report for given run date
-    # logging.info(f'Running Qube BOS By Block report for {qube_date}')
+    # logger.info(f'Running Qube BOS By Block report for {qube_date}')
     qube_by_block_sql = join_sql_queries(
         QUBE_BOS_SHEET_BY_BLOCK_SQL,
         QUBE_BOS_REPORT_BY_BLOCK_SQL,
         BOS_ACCOUNT_BALANCES_BY_BLOCK_SQL,
     )
-    # logging.debug(qube_by_block_sql)
-    # qube_by_block_df = run_sql_query(db_conn, qube_by_block_sql, (qube_date, qube_date, bos_date))
+    # logger.debug(qube_by_block_sql)
+    # qube_by_block_df = run_sql_query(db_conn, qube_by_block_sql, (qube_date, qube_date, bos_date), logger)
     # qube_by_block_df = add_extra_rows(qube_by_block_df)
     # qube_by_block_df = qube_by_block_df.sort_values(by='Property / Block')
     # qube_by_block_df = add_column_totals(qube_by_block_df)
@@ -535,27 +445,28 @@ def runReports(db_conn: sqlite3.Connection, qube_date: str, bos_date: str) -> No
     # qube_by_block_df.to_excel(excel_writer, sheet_name='Qube BOS By Block {}'.format(qube_date), index=False, float_format = '%.2f')
 
     # Run Qube BOS By Property report for given run date
-    # logging.info(f'Running Qube BOS By Property report for {qube_date}')
+    # logger.info(f'Running Qube BOS By Property report for {qube_date}')
     qube_by_property_sql = join_sql_queries(
         QUBE_BOS_SHEET_BY_PROPERTY_SQL,
         QUBE_BOS_REPORT_BY_PROPERTY_SQL,
         BOS_ACCOUNT_BALANCES_BY_PROPERTY_SQL,
     )
-    # logging.debug(qube_by_property_sql)
-    # qube_by_property_df = run_sql_query(db_conn, qube_by_property_sql, (qube_date, qube_date, bos_date))
+    # logger.debug(qube_by_property_sql)
+    # qube_by_property_df = run_sql_query(db_conn, qube_by_property_sql, (qube_date, qube_date, bos_date), logger)
     # qube_by_property_df = qube_by_property_df.sort_values(by='Property / Block')
     # qube_by_property_df = add_column_totals(qube_by_property_df)
     # qube_by_property_df.drop(['BOS GR', 'Discrepancy GR'], axis=1, inplace=True)
     # qube_by_property_df.to_excel(excel_writer, sheet_name='Qube BOS By Property {}'.format(qube_date), index=False, float_format = '%.2f')
 
     # Run Qube BOS report for given run date
-    logging.info(f"Running Qube BOS report for {qube_date}")
+    logger.info(f"Running Qube BOS report for {qube_date}")
     sql = union_sql_queries(qube_by_property_sql, qube_by_block_sql)
-    logging.debug(sql)
+    logger.debug(sql)
     df = run_sql_query(
         db_conn,
         sql,
         (qube_date, qube_date, bos_date) + (qube_date, qube_date, bos_date),
+        logger
     )
     df = add_extra_rows(df)
     df = df.sort_values(by="Property / Block")
@@ -569,9 +480,9 @@ def runReports(db_conn: sqlite3.Connection, qube_date: str, bos_date: str) -> No
     )
 
     # Run SC total transactions by tenant report for given run date
-    logging.info(f"Running Total SC Paid By Tenant on {bos_date} report")
-    logging.debug(SELECT_TOTAL_PAID_SC_BY_TENANT_SQL)
-    df = run_sql_query(db_conn, SELECT_TOTAL_PAID_SC_BY_TENANT_SQL, (bos_date,) * 2)
+    logger.info(f"Running Total SC Paid By Tenant on {bos_date} report")
+    logger.debug(SELECT_TOTAL_PAID_SC_BY_TENANT_SQL)
+    df = run_sql_query(db_conn, SELECT_TOTAL_PAID_SC_BY_TENANT_SQL, (bos_date,) * 2, logger)
     df = add_column_totals(df)
     df.to_excel(
         excel_writer,
@@ -624,19 +535,19 @@ def main(qube_date: dt.date = None, bos_date: dt.date = None) -> None:
 
     os.makedirs(WPP_REPORT_DIR, exist_ok=True)
 
-    logging.info("Running Reports")
+    logger.info("Running Reports")
     try:
         db_conn = sqlite3.connect(WPP_DB_FILE)
         qube_date, bos_date = get_run_date_args(args, qube_date, bos_date)
         runReports(db_conn, qube_date, bos_date)
     except Exception as ex:
-        logging.error(str(ex))
+        logger.error(str(ex))
 
     elapsed_time = time.time() - start_time
     time.strftime("%S", time.gmtime(elapsed_time))
 
-    logging.info("Done in {} seconds.".format(round(elapsed_time, 1)))
-    logging.info(
+    logger.info("Done in {} seconds.".format(round(elapsed_time, 1)))
+    logger.info(
         "----------------------------------------------------------------------------------------"
     )
     # input("Press enter to end.")
